fixed routing issues with registration, post and delete functions, fixed Link button styling, added new post link to profile page, Fixed navbar profile route issue
