+++ photography-connect-client/src/App.css
@@ -83,8 +83,9 @@ body{
 
 .profile-image{
   margin:10px;
-  padding:10px;
-  width:400px;
+  padding:5px;
+  max-width:95%;
+  max-height:400px;
   border-radius:50%;
   background-color: #706677;
   display:block;
@@ -133,6 +134,7 @@ body{
   margin-right: auto;
   border-radius: 5px;
   margin-bottom:10px;
+  max-height:400px;
 }
 
 @keyframes fade-in{
@@ -159,4 +161,23 @@ body{
 
 .modal-footer {
   border-top: 0 none;
+}
+
+.error{
+  color:red;
+  background-color:rgb(255, 255, 0);
+}
+
+.new-post-button{
+  font-size: 30px;
+  margin-top:35px;
+  display:block;
+  margin-left:auto;
+  margin-right:auto;
+  width:80%;
+  height: 100px;
+}
+
+.link{
+  text-decoration: none;
 }
