+++ photography-connect-client/src/components/LoggedInHome.js
@@ -1,12 +1,11 @@
 import React from "react";
 import "../App.css";
-import { Link } from "react-router-dom";
 import PostAPI from "./PostAPI";
 
 function LoggedInHome(){
     return(
-        <div className="">
-            <h1>You're logged in!</h1>
+        <div>
+            <h1  className="centered">You're logged in!</h1>
             <PostAPI/>
         </div>
     )
