+++ photography-connect-client/src/views/Profile.js
@@ -7,7 +7,7 @@ import {Link, useParams} from "react-router-dom";
 import UserPostAPI from "../components/UserPostAPI";
 import jwt_decode from "jwt-decode";
 
-function Profile(props){
+function Profile(){
     const params = useParams();
     const token=sessionStorage.getItem("AuthToken")
     const decodedToken= jwt_decode(token);
@@ -28,7 +28,7 @@ function Profile(props){
             // setSelectedUser1(decodedToken.user_id)
             // console.log(params.userId)
             // setSelectedUser1(params.userId)
-            console.log(selectedUser1)
+            // console.log(selectedUser1)
             const config={
                  headers:{
                     Authorization : `${token}` 
@@ -48,200 +48,364 @@ function Profile(props){
             })
         }
         getUserDetails()
-        },[])
+        },[params.userId])
 
-    if(instaLink !== "" && facebookLink === ""){
-        return(
-            <div className="container-fluid profile-page">
-                <div className="row">
-                    <div className="col-lg-5 col-sm-12">
-                        <div className="bio-card">
-                            <img src={userImage} className="img-fluid profile-image mx-auto" alt="" />                        
+    if(params.userId === decodedToken.user_id){
+        if(instaLink !== "" && facebookLink === ""){
+            return(
+                <div className="container-fluid profile-page">
+                    <div className="row">
+                        <div className="col-lg-5 col-sm-12">
+                            <div className="bio-card">
+                                <img src={userImage} className="img-fluid profile-image mx-auto" alt="" />                        
+                            </div>
+                        </div>
+                        <div className="col-lg-7 col-sm-12">
+                            <div className="bio-card">
+                                <h1 className="bio-username">{firstName} {lastName}</h1>
+                                <p>
+                                    Occupation: {occupation}
+                                </p>
+                                <p>
+                                    Bio: {bio}
+                                </p>
+                                <p>
+                                    Social Links:
+                                    <a className="social-link" href={instaLink} target="_blank" rel="noreferrer noopener">
+                                        <img src={InstaImage} className="img-fluid mx-auto" width="30px" alt="Instagram Link" />                                
+                                    </a>                            
+                                </p>
+                                <Link to={`/profile/edit/${params.userId}`}>
+                                    <button className="btn">
+                                        Edit Bio
+                                    </button>
+                                </Link>
+                                <Link to="/profile/uploadprofileimage">
+                                    <button className="btn">
+                                        Upload Profile Image
+                                    </button>
+                                </Link>
+                            </div>
+        
                         </div>
                     </div>
-                    <div className="col-lg-7 col-sm-12">
-                        <div className="bio-card">
-                            <h1 className="bio-username">{firstName} {lastName}</h1>
-                            <p>
-                                Occupation: {occupation}
-                            </p>
-                            <p>
-                                Bio: {bio}
-                            </p>
-                            <p>
-                                Social Links:
-                                <a className="social-link" href={instaLink} target="_blank" rel="noreferrer noopener">
-                                    <img src={InstaImage} className="img-fluid mx-auto" width="30px" alt="Instagram Link" />                                
-                                </a>                            
-                            </p>
-                            <button className="btn">
-                                <Link to={`/profile/edit/${params.userId}`}>
-                                    Edit Bio
-                                </Link>
-                            </button>
-                            <button className="btn">
-                                <Link to="/profile/uploadprofileimage">
-                                    Upload Profile Image
-                                </Link>
-                            </button>
+                    <div className="row profile-posts">
+                        <div className="col-lg-12 col-sm-12">
+                            <UserPostAPI selectedUser1={selectedUser1} currentUserId={decodedToken.user_id}/>
+                            
                         </div>
-    
                     </div>
                 </div>
-                <div className="row profile-posts">
-                    <div className="col-lg-12 col-sm-12">
+        
+            )
+        }
+        else if(instaLink === "" && facebookLink !== ""){
+            return(
+                <div className="container-fluid profile-page">
+                    <div className="row">
+                        <div className="col-lg-5 col-sm-12">
+                            <div className="bio-card">
+                                <img src={userImage} className="img-fluid profile-image mx-auto" alt="" />                        
+                            </div>
+                        </div>
+                        <div className="col-lg-7 col-sm-12">
+                            <div className="bio-card">
+                                <h1 className="bio-username">{firstName} {lastName}</h1>
+                                <p>
+                                    Occupation: {occupation}
+                                </p>
+                                <p>
+                                    Bio: {bio}
+                                </p>
+                                <p>
+                                    Social Links:
+                                    <a className="social-link" href={facebookLink} target="_blank" rel="noreferrer noopener">
+                                        <img src={FacebookImage} className="img-fluid mx-auto" width="30px" alt="Facebook Link" />                                
+                                    </a>                          
+                                </p>
+                                <Link to={`/profile/edit/${params.userId}`}>
+                                    <button className="btn">
+                                        Edit Bio
+                                    </button>
+                                </Link>
+                                <Link to="/profile/uploadprofileimage">
+                                <button className="btn">
+                                    Upload Profile Image
+                                </button>
+                                </Link>
+                            </div>
+        
+                        </div>
+                    </div>
+                    <div className="row profile-posts">
+                        <div className="col-lg-12 col-sm-12">
                         <UserPostAPI selectedUser1={selectedUser1}/>
-                        
-                    </div>
-                </div>
-            </div>
-    
-        )
-    }
-    else if(instaLink === "" && facebookLink !== ""){
-        return(
-            <div className="container-fluid profile-page">
-                <div className="row">
-                    <div className="col-lg-5 col-sm-12">
-                        <div className="bio-card">
-                            <img src={userImage} className="img-fluid profile-image mx-auto" alt="" />                        
+                            
                         </div>
                     </div>
-                    <div className="col-lg-7 col-sm-12">
-                        <div className="bio-card">
-                            <h1 className="bio-username">{firstName} {lastName}</h1>
-                            <p>
-                                Occupation: {occupation}
-                            </p>
-                            <p>
-                                Bio: {bio}
-                            </p>
-                            <p>
-                                Social Links:
-                                <a className="social-link" href={facebookLink} target="_blank" rel="noreferrer noopener">
-                                    <img src={FacebookImage} className="img-fluid mx-auto" width="30px" alt="Facebook Link" />                                
-                                </a>                          
-                            </p>
-                            <button className="btn">
+                </div>
+        
+            )
+        }
+        else if(instaLink === "" && facebookLink === ""){
+            return(
+                <div className="container-fluid profile-page">
+                    <div className="row">
+                        <div className="col-lg-5 col-sm-12">
+                            <div className="bio-card">
+                                <img src={userImage} className="img-fluid profile-image mx-auto" alt="" />                        
+                            </div>
+                        </div>
+                        <div className="col-lg-7 col-sm-12">
+                            <div className="bio-card">
+                                <h1 className="bio-username">{firstName} {lastName}</h1>
+                                <p>
+                                    Occupation: {occupation}
+                                </p>
+                                <p>
+                                    Bio: {bio}
+                                </p>
+                                <p>
+                                    Social Links: No social links to display                         
+                                </p>
                                 <Link to={`/profile/edit/${params.userId}`}>
-                                    Edit Bio
+                                    <button className="btn">
+                                        Edit Bio
+                                    </button>
                                 </Link>
-                            </button>
-                            <button className="btn">
                                 <Link to="/profile/uploadprofileimage">
+                                <button className="btn">
                                     Upload Profile Image
+                                </button>
                                 </Link>
-                            </button>
-                        </div>
-    
-                    </div>
-                </div>
-                <div className="row profile-posts">
-                    <div className="col-lg-12 col-sm-12">
-                    <UserPostAPI selectedUser1={selectedUser1}/>
-                        
-                    </div>
-                </div>
-            </div>
-    
-        )
-    }
-    else if(instaLink === "" && facebookLink === ""){
-        return(
-            <div className="container-fluid profile-page">
-                <div className="row">
-                    <div className="col-lg-5 col-sm-12">
-                        <div className="bio-card">
-                            <img src={userImage} className="img-fluid profile-image mx-auto" alt="" />                        
+                            </div>
+        
                         </div>
                     </div>
-                    <div className="col-lg-7 col-sm-12">
-                        <div className="bio-card">
-                            <h1 className="bio-username">{firstName} {lastName}</h1>
-                            <p>
-                                Occupation: {occupation}
-                            </p>
-                            <p>
-                                Bio: {bio}
-                            </p>
-                            <p>
-                                Social Links: No social links to display                         
-                            </p>
-                            <button className="btn">
-                                <Link to={`/profile/edit/${params.userId}`}>
-                                    Edit Bio
-                                </Link>
-                            </button>
-                            <button className="btn">
-                                <Link to="/profile/uploadprofileimage">
-                                    Upload Profile Image
-                                </Link>
-                            </button>
+                    <div className="row">
+                        <div className="col-lg-12 col-sm-12">
+                            <Link className="link" to={`/newpost`}>
+                                <button className="btn new-post-button">New Post</button>
+                            </Link>
                         </div>
-    
+                    </div>
+                    <div className="row profile-posts">
+                        <div className="col-lg-12 col-sm-12">
+                        <UserPostAPI selectedUser1={selectedUser1}/>
+                            
+                        </div>
                     </div>
                 </div>
-                <div className="row profile-posts">
-                    <div className="col-lg-12 col-sm-12">
-                    <UserPostAPI selectedUser1={selectedUser1}/>
-                        
+        
+            )
+        }
+        else{
+            return(
+                <div className="container-fluid profile-page">
+                    <div className="row">
+                        <div className="col-lg-5 col-sm-12">
+                            <div className="bio-card">
+                                <img src={userImage} className="img-fluid profile-image mx-auto" alt="" />                        
+                            </div>
+                        </div>
+                        <div className="col-lg-7 col-sm-12">
+                            <div className="bio-card">
+                                <h1 className="bio-username">{firstName} {lastName}</h1>
+                                <p>
+                                    Occupation: {occupation}
+                                </p>
+                                <p>
+                                    Bio: {bio}
+                                </p>
+                                <p>
+                                    Social Links:
+                                    <a className="social-link" href={facebookLink} target="_blank" rel="noreferrer noopener">
+                                        <img src={FacebookImage} className="img-fluid mx-auto" width="30px" alt="Facebook Link" />                                
+                                    </a>
+                                    <a className="social-link" href={instaLink} target="_blank" rel="noreferrer noopener">
+                                        <img src={InstaImage} className="img-fluid mx-auto" width="30px" alt="Instagram Link" />                                
+                                    </a>                            
+                                </p>
+                                    <Link to={`/profile/edit/${params.userId}`}>
+                                        <button className="btn">
+                                            Edit Bio
+                                        </button>
+                                    </Link>
+                                <button className="btn">
+                                    <Link to="/profile/uploadprofileimage">
+                                        Upload Profile Image
+                                    </Link>
+                                </button>
+                            </div>
+        
+                        </div>
+                    </div>
+                    <div className="row profile-posts">
+                        <div className="col-lg-12 col-sm-12">
+                            <UserPostAPI/>
+                            
+                        </div>
                     </div>
                 </div>
-            </div>
-    
-        )
+        
+            )
+        }
     }
     else{
-        return(
-            <div className="container-fluid profile-page">
-                <div className="row">
-                    <div className="col-lg-5 col-sm-12">
-                        <div className="bio-card">
-                            <img src={userImage} className="img-fluid profile-image mx-auto" alt="" />                        
+        if(instaLink !== "" && facebookLink === ""){
+            return(
+                <div className="container-fluid profile-page">
+                    <div className="row">
+                        <div className="col-lg-5 col-sm-12">
+                            <div className="bio-card">
+                                <img src={userImage} className="img-fluid profile-image mx-auto" alt="" />                        
+                            </div>
+                        </div>
+                        <div className="col-lg-7 col-sm-12">
+                            <div className="bio-card">
+                                <h1 className="bio-username">{firstName} {lastName}</h1>
+                                <p>
+                                    Occupation: {occupation}
+                                </p>
+                                <p>
+                                    Bio: {bio}
+                                </p>
+                                <p>
+                                    Social Links:
+                                    <a className="social-link" href={instaLink} target="_blank" rel="noreferrer noopener">
+                                        <img src={InstaImage} className="img-fluid mx-auto" width="30px" alt="Instagram Link" />                                
+                                    </a>                            
+                                </p>
+                            </div>
+        
                         </div>
                     </div>
-                    <div className="col-lg-7 col-sm-12">
-                        <div className="bio-card">
-                            <h1 className="bio-username">{firstName} {lastName}</h1>
-                            <p>
-                                Occupation: {occupation}
-                            </p>
-                            <p>
-                                Bio: {bio}
-                            </p>
-                            <p>
-                                Social Links:
-                                <a className="social-link" href={facebookLink} target="_blank" rel="noreferrer noopener">
-                                    <img src={FacebookImage} className="img-fluid mx-auto" width="30px" alt="Facebook Link" />                                
-                                </a>
-                                <a className="social-link" href={instaLink} target="_blank" rel="noreferrer noopener">
-                                    <img src={InstaImage} className="img-fluid mx-auto" width="30px" alt="Instagram Link" />                                
-                                </a>                            
-                            </p>
-                            <button className="btn">
-                                <Link to={`/profile/edit/${params.userId}`}>
-                                    Edit Bio
-                                </Link>
-                            </button>
-                            <button className="btn">
-                                <Link to="/profile/uploadprofileimage">
-                                    Upload Profile Image
-                                </Link>
-                            </button>
+                    <div className="row profile-posts">
+                        <div className="col-lg-12 col-sm-12">
+                            <UserPostAPI selectedUser1={selectedUser1}/>
+                            
                         </div>
-    
                     </div>
                 </div>
-                <div className="row profile-posts">
-                    <div className="col-lg-12 col-sm-12">
-                        <UserPostAPI/>
-                        
+        
+            )
+        }
+        else if(instaLink === "" && facebookLink !== ""){
+            return(
+                <div className="container-fluid profile-page">
+                    <div className="row">
+                        <div className="col-lg-5 col-sm-12">
+                            <div className="bio-card">
+                                <img src={userImage} className="img-fluid profile-image mx-auto" alt="" />                        
+                            </div>
+                        </div>
+                        <div className="col-lg-7 col-sm-12">
+                            <div className="bio-card">
+                                <h1 className="bio-username">{firstName} {lastName}</h1>
+                                <p>
+                                    Occupation: {occupation}
+                                </p>
+                                <p>
+                                    Bio: {bio}
+                                </p>
+                                <p>
+                                    Social Links:
+                                    <a className="social-link" href={facebookLink} target="_blank" rel="noreferrer noopener">
+                                        <img src={FacebookImage} className="img-fluid mx-auto" width="30px" alt="Facebook Link" />                                
+                                    </a>                          
+                                </p>
+                            </div>
+        
+                        </div>
+                    </div>
+                    <div className="row profile-posts">
+                        <div className="col-lg-12 col-sm-12">
+                        <UserPostAPI selectedUser1={selectedUser1}/>
+                            
+                        </div>
                     </div>
                 </div>
-            </div>
-    
-        )
+        
+            )
+        }
+        else if(instaLink === "" && facebookLink === ""){
+            return(
+                <div className="container-fluid profile-page">
+                    <div className="row">
+                        <div className="col-lg-5 col-sm-12">
+                            <div className="bio-card">
+                                <img src={userImage} className="img-fluid profile-image mx-auto" alt="" />                        
+                            </div>
+                        </div>
+                        <div className="col-lg-7 col-sm-12">
+                            <div className="bio-card">
+                                <h1 className="bio-username">{firstName} {lastName}</h1>
+                                <p>
+                                    Occupation: {occupation}
+                                </p>
+                                <p>
+                                    Bio: {bio}
+                                </p>
+                                <p>
+                                    Social Links: No social links to display                         
+                                </p>
+                            </div>
+        
+                        </div>
+                    </div>
+                    <div className="row profile-posts">
+                        <div className="col-lg-12 col-sm-12">
+                        <UserPostAPI selectedUser1={selectedUser1}/>
+                            
+                        </div>
+                    </div>
+                </div>
+        
+            )
+        }
+        else{
+            return(
+                <div className="container-fluid profile-page">
+                    <div className="row">
+                        <div className="col-lg-5 col-sm-12">
+                            <div className="bio-card">
+                                <img src={userImage} className="img-fluid profile-image mx-auto" alt="" />                        
+                            </div>
+                        </div>
+                        <div className="col-lg-7 col-sm-12">
+                            <div className="bio-card">
+                                <h1 className="bio-username">{firstName} {lastName}</h1>
+                                <p>
+                                    Occupation: {occupation}
+                                </p>
+                                <p>
+                                    Bio: {bio}
+                                </p>
+                                <p>
+                                    Social Links:
+                                    <a className="social-link" href={facebookLink} target="_blank" rel="noreferrer noopener">
+                                        <img src={FacebookImage} className="img-fluid mx-auto" width="30px" alt="Facebook Link" />                                
+                                    </a>
+                                    <a className="social-link" href={instaLink} target="_blank" rel="noreferrer noopener">
+                                        <img src={InstaImage} className="img-fluid mx-auto" width="30px" alt="Instagram Link" />                                
+                                    </a>                            
+                                </p>
+                            </div>
+        
+                        </div>
+                    </div>
+                    <div className="row profile-posts">
+                        <div className="col-lg-12 col-sm-12">
+                            <UserPostAPI/>
+                            
+                        </div>
+                    </div>
+                </div>
+        
+            )
+        }
     }
+
 
 }
